Add RBAC for Test Repositories : 105
* Migrated database to include new team/test repo role mappings
* Added Test Repo permission checks in update/delete repo
* Added code to add test repo/team role mappings to db - Currently broken
* Updated the calls to crud in routers to include requisite user object
* Updated roles and actions to include update/delete roles/actions
* Added user_product fixture in tests/test_api/test_repositories.py to
  ensure there's always a team created to test with

Resolves: https://github.com/AlmaLinux/build-system/issues/105

+++ alws/crud/test_repository.py
@@ -2,12 +2,35 @@ import typing
 
 from sqlalchemy import delete, or_, select
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.orm import joinedload
+from sqlalchemy.orm import joinedload, selectinload
 from sqlalchemy.sql.expression import func
 
 from alws import models
+from alws.crud.teams import get_teams
 from alws.errors import DataNotFoundError, TestRepositoryError
 from alws.schemas import test_repository_schema
+from alws.perms.authorization import can_perform
+from alws.errors import (
+    PermissionDenied,
+)
+from alws.perms import actions
+from alws.crud import user as user_crud
+from alws.perms.roles import (
+    Contributor,
+    Observer,
+)
+from alws.perms.actions import (
+    UpdateTest,
+    DeleteTest,
+)
+from alws.models import (
+    Team,
+    User,
+    UserAction,
+    UserRole,
+)
+from alws.crud.actions import ensure_all_actions_exist
+
 
 
 async def get_repositories(
@@ -24,7 +47,12 @@ async def get_repositories(
         query = (
             select(models.TestRepository)
             .order_by(models.TestRepository.id.desc())
-            .options(joinedload(models.TestRepository.packages))
+            .options(
+                joinedload(models.TestRepository.packages),
+                joinedload(models.TestRepository.team).joinedload(models.Team.roles),
+                joinedload(models.TestRepository.owner),
+                joinedload(models.TestRepository.roles),
+            )
         )
         if name:
             query = query.where(models.TestRepository.name == name)
@@ -139,6 +167,41 @@ async def bulk_create_package_mapping(
     await session.flush()
 
 
+def get_team_role_name(team_name: str, role_name: str):
+    return f'{team_name}_{role_name}'
+
+async def create_test_repository_role_mapping(session: AsyncSession, team_name: str):
+    required_roles = (Contributor,)
+    required_actions = (UpdateTest, DeleteTest)
+    role_names = [get_team_role_name(team_name, role.name) for role in required_roles]
+
+    existing_roles = (
+        await session.execute(
+            select(UserRole).where(
+                UserRole.name.in_(role_names)
+            ).options(
+                selectinload(UserRole.actions)
+            )
+        )
+    ).scalars().all()
+    await ensure_all_actions_exist(session)
+    existing_actions = (await session.execute(
+        select(UserAction))).scalars().all()
+    actions_to_add = []
+    for existing_action in existing_actions:
+        for action in required_actions:
+            if action.name == existing_action.name:
+                actions_to_add.append(existing_action)
+
+    for role in existing_roles:
+        for action in actions_to_add:
+            if action not in role.actions:
+                role.actions.append(action)
+    await session.flush()
+
+    return existing_roles
+
+
 async def create_repository(
     session: AsyncSession,
     payload: test_repository_schema.TestRepositoryCreate,
@@ -163,6 +226,12 @@ async def create_repository(
         raise TestRepositoryError("Test Repository already exists")
 
     repository = models.TestRepository(**payload.model_dump())
+
+    team = await get_teams(session, team_id=payload.team_id)
+    repository_roles = await create_test_repository_role_mapping(session, team.name)
+    repository.team = team
+    repository.roles = repository_roles
+
     session.add(repository)
     await session.flush()
     await session.refresh(repository)
@@ -173,6 +242,7 @@ async def update_repository(
     session: AsyncSession,
     repository_id: int,
     payload: test_repository_schema.TestRepositoryUpdate,
+    user: models.User,
 ) -> models.TestRepository:
     db_repo = await get_repositories(
         session,
@@ -180,6 +250,12 @@ async def update_repository(
     )
     if not db_repo:
         raise DataNotFoundError(f"Unknown test repository ID: {repository_id}")
+
+    db_user = await user_crud.get_user(session, user_id=user.id)
+    if not can_perform(db_repo, db_user, actions.UpdateTest.name):
+        raise PermissionDenied(
+            "User does not have permissions to update this test repository"
+        )
 
     for field, value in payload.model_dump().items():
         setattr(db_repo, field, value)
@@ -211,11 +287,21 @@ async def bulk_delete_package_mapping(
     await session.flush()
 
 
-async def delete_repository(session: AsyncSession, repository_id: int):
+async def delete_repository(
+        session: AsyncSession,
+        repository_id: int,
+        user: models.User,
+):
     db_repo = await get_repositories(session, repository_id=repository_id)
     if not db_repo:
         raise DataNotFoundError(
             f"Test repository={repository_id} doesn`t exist"
+        )
+
+    db_user = await user_crud.get_user(session, user_id=user.id)
+    if not can_perform(db_repo, db_user, actions.DeleteTest.name):
+        raise PermissionDenied(
+            "User does not have permissions to delete this test repository"
         )
     await session.delete(db_repo)
     await session.flush()
